pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SciFiReaders"
dynamic = ["version"]
description = "Tools for extracting data and metadata from scientific data files"
readme = "README.rst"
license = { text = "MIT" }
authors = [
    { name = "Pycroscopy contributors", email = "pycroscopy@gmail.com" }
]
keywords = ["imaging", "spectra", "multidimensional", "scientific"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering :: Information Analysis"
]
requires-python = ">=3.9"
dependencies = [
    "setuptools",
    "numpy",
    "toolz",
    "cytoolz",
    "dask>=2.20.0",
    "sidpy-lite",
    "ipython>=7.1.0",
    "gdown",
]

[project.optional-dependencies]
hyperspy = ["hyperspy"]
igor2 = ["igor2"]
gwyddion = ["gwyfile"]
image = ["pillow", "tifffile"]
test = ["pytest", "hyperspy",  "gwyfile", "gdown"]

[project.urls]
Homepage = "https://pycroscopy.github.io/SciFiReaders/about.html"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["*.tests", "*.tests.*", "tests.*", "tests"]

[tool.setuptools.dynamic]
version = { attr = "SciFiReaders.__version__.version" }
